+++ Tuning/FFTaux.py
@@ -1,7 +1,7 @@
 """
 auxiliary functions
 """
-from __future__ import annotations
+# from __future__ import annotations
 
 import logging
 from functools import wraps