+++ Tuning/L1costfunction.py
@@ -1,8 +1,10 @@
 """
-minimizer for L1 norm
+cost function of the L1 norm, to be minimized. L1 being the sum of the absolute
+deviations of the calculated resonance frequencies from the measured ones,
+normalized to the measured frequencies.
 """
 
-from __future__ import annotations
+# from __future__ import annotations
 
 from numdifftools import Jacobian, Hessian
 from numpy import sqrt, array, sign, nan
@@ -13,11 +15,11 @@ class L1(object):
     def __init__(
             self,
             ind: list[tuple]
-    ):
+    ) -> None:
         """
         :param ind: list of tuples -
         1st element of tuple: measured resonance frequencies as from peaks (FFT)
-        2nd element of tuple: its partial
+        2nd element of tuple: its partial i
         """
         self.__fo = ind
         self.l1_first: float = nan
@@ -32,13 +34,14 @@ class L1(object):
         """
         returns the cost function for a regression on the L1 norm
         l1 = sum( abs( f_i(measured) - f_i(calculated) ) / f_i(measured) )
-        :param x0: NDArray - [f0, b] such that
+        :param x0 - [f0, b] provided by minimizer, such that
             f = i * x0[0] * sqrt(1. + x0[1] * i**2)
             inharmonicity of a vibrating string for partial i
-        :param jac: bool - if jacobi is to be calculated
-        :return: float - l1 cost function
+        :param jac - if jacobi is to be calculated
+        :return - l1 cost function
         other property:
-        self.jacobi - NDArray - derivatives δl1/δf_0 and δl1/δB
+        if jac is True the Jacobian is calculated analytically and stored in
+        self.jacobi - derivatives δl1/δf_0 and δl1/δB
         """
         l1 = 0.  # l1 cost function
         self.jacobi = array([0., 0.])
@@ -46,7 +49,7 @@ class L1(object):
         for found in self.__fo:
             f_calc = found[1] * x0[0] * sqrt(1. + x0[1] * found[1] ** 2)
             diff = f_calc - found[0]
-            # L1 norm normalized to frequency, as L1 increases from bass to discant
+            # L1 norm normalized to frequency, as L1 varies with frequency
             l1 += abs(diff) / found[0]
             if jac:
                 # n-th partial is index + 1
@@ -55,8 +58,7 @@ class L1(object):
                     i=found[1],
                     trova=found[0]
                 ) * sign(diff)
-        if self.l1_first is nan:
-            self.l1_first = l1
+        if self.l1_first is nan: self.l1_first = l1
         self.l1_last = l1
 
         return l1
@@ -66,24 +68,50 @@ class L1(object):
             x0: NDArray,
             jac: bool = False
     ) -> float:
-        # b is coming in as log10(b)
+        """
+        b is coming in as log10(b), used with bruteforce. Not used with SLSQP
+        :param x0:
+        :param jac:
+        :return:
+        """
         x0[1] = 10 ** x0[1]
         return self.l1_minimum(x0, jac)
 
     def l1_minimum_jac_direct(self, x0: NDArray) -> NDArray:
+        """
+        not used, but kept for reference
+        :param x0:
+        :return:
+        """
         self.l1_minimum(x0, jac=True)
         return self.jacobi
 
     def l1_minimum_jac(self, x0: NDArray) -> NDArray:
+        """
+        Calculate Jacobian with finite difference approximation
+        :param x0:
+        :return:
+        """
         return Jacobian(self.l1_minimum(x0))(x0).ravel()
 
     def l1_minimum_hess(self, x0: NDArray) -> NDArray:
+        """
+        Calculate Hessian with finite difference approximation
+        :param x0:
+        :return:
+        """
         return Hessian(self.l1_minimum(x0))(x0)
 
     def l1_minimum_der(self, x0: NDArray) -> tuple[float, NDArray]:
+        """
+        returns both the L1 cost function and its derivatives, computed analytically
+        :param x0:
+        :return:
+        """
         return self.l1_minimum(x0, jac=True), self.jacobi
 
     def compare_l1(self) -> bool:
+        """compare first and last L1 cost function calculation."""
         return self.l1_last < self.l1_first
 
     @staticmethod
@@ -92,8 +120,17 @@ class L1(object):
             i: int,
             trova: float
     ) -> NDArray:
-        x0[1] = max(0., x0[1])
+        """
+        computes analytically the normalized derivatives of the L1 cost
+        function with respect to f0 and b for an individual partial i
+        :param x0:
+        :param i: partial number
+        :param trova: measured frequency at partial i
+        :return:
+        """
+        x0[1] = max(0., x0[1])  # let b be non-negative
         tmp = sqrt(1. + x0[1] * i ** 2)
+
         # derivative with respect to base frequency
         deriv_f0 = i * tmp / trova
         # derivative with respect to inharmonicity